moved functions from project.js to projects.js
