+++ src/utils/data/projects.js
@@ -3,6 +3,8 @@ import {
 	collection,
 	getDocs,
 	getDoc,
+	updateDoc,
+	arrayUnion,
 	doc,
 	query,
 	where,
@@ -59,4 +61,17 @@ export const getProjectByInvolvedId = async (id) => {
 	});
 	return result;
 };
+
+export const getImageUrls = async (projectId) => {
+	const project = await getDoc(doc(db, 'projects', projectId));
+	return project.data().imageUrls;
+};
+
+export const addImageUrl = async (projectId, imageUrl) => {
+	await updateDoc(doc(db, 'projects', projectId), {
+		imageUrls: arrayUnion(imageUrl),
+	});
+	return getImageUrls(projectId);
+};
+
 // Create service request is the functionality to be used for creating a project
